Rename SystemTestPartitionsTestHelper
